+++ public/data/feeds.json
@@ -1,0 +1,5 @@
+[
+  { "id": "a", "nickname": "abc" },
+  { "id": "b", "nickname": "abcd" },
+  { "id": "c", "nickname": "abce" }
+]
