+++ src/pages/kimminki/Main/Main.scss
@@ -94,10 +94,16 @@
     margin-top: 100px;
     padding-top: 75px;
 
+    .feedss {
+      display: flex;
+      flex-direction: column;
+    }
+
     .feeds {
       position: relative;
       right: -240px;
       width: 600px;
+      margin-bottom: 70px;
       padding-top: 15px;
       padding-right: 10px;
       padding-left: 10px;
